functions/requirements.txt
# Azure Functions Python v2 - Requirements
azure-functions>=1.17.0
azure-identity>=1.15.0
azure-mgmt-costmanagement>=4.0.0
azure-mgmt-compute>=30.0.0
azure-mgmt-network>=25.0.0
azure-mgmt-resource>=23.0.0
azure-mgmt-advisor>=9.0.0
azure-mgmt-monitor>=6.0.0
azure-data-tables>=12.5.0
requests>=2.31.0
python-dateutil>=2.8.2
